SH-12 Logic for language dropdown
