+++ src/shared/components/header/header.component.html
@@ -1,3 +1,5 @@
 <app-top-header></app-top-header>
-<app-middle-header class="navbar"></app-middle-header>
-<app-sub-header  class="header"></app-sub-header>
+
+<app-middle-header></app-middle-header>
+
+<app-sub-header></app-sub-header>
