--- src/shared/components/header/index.ts
@@ -1,1 +1,0 @@
-export * from './header.component';
