--- src/shared/components/header/middle-header/index.ts
@@ -1,1 +1,0 @@
-export * from './middle-header.component';
