--- src/shared/components/header/sub-header/index.ts
@@ -1,1 +1,0 @@
-export * from './sub-header.component';
