--- src/shared/components/header/top-header/index.ts
@@ -1,1 +1,0 @@
-export * from './top-header.component';
