+++ src/shared/components/index.ts
@@ -1,7 +1,7 @@
 export * from './header/header.component';
-export * from './header/top-header/top-header.component';
-export * from './header/middle-header/middle-header.component';
-export * from './header/sub-header/sub-header.component';
+export * from './top-header/top-header.component';
+export * from './middle-header/middle-header.component';
+export * from './sub-header/sub-header.component';
 export * from './searchbar/searchbar.component';
 export * from './loader/loader.component';
 export * from './cart-icon/cart-icon.component';
