+++ src/shared/components/searchbar/searchbar.component.html
@@ -2,6 +2,6 @@
       class="rounded-2xl border-2 border-solid border-b-blue-600 text-blue-600 appearance-none py-2.5 px-5 w-full focus:border-b-blue-500 focus:outline-0"
       type="text"
       [formControl]="searchValue"
-      [placeholder]="'sharedComponents.middle-header.searchForProducts' | translate">
+      [placeholder]="'sharedComponents.navbar.searchForProducts' | translate">
 
     <button *ngIf="searchValue.value" class="absolute right-3.5 w-5 h-5 rounded-full text-center bg-red-400" (click)="deleteSearch()">X</button>
