+++ src/shared/components/top-header/top-header.component.html
@@ -1,0 +1,7 @@
+<div class="flex flex-row px-40 py-2 bg-grey-100 text-grey-600">
+    <svg-icon [src]="'localization-icon'"></svg-icon>
+  <select (change)="onChange($event)">
+    <option disabled selected>Language</option>
+    <option *ngFor="let option of dropdownOptions" [value]=option.value >{{ option.label }}</option>
+  </select>
+</div>
